backend/requirements.txt
django>=6.0.2
djangorestframework>=3.16.1
pymysql>=1.1.2
